+++ packages/flame/lib/src/components/component.dart
@@ -131,7 +131,7 @@ class Component with Loadable {
   void render(Canvas canvas) {}
 
   void renderTree(Canvas canvas) {
-    preRender(canvas);
+    preRender(canvas); // ignore: deprecated_member_use_from_same_package
     render(canvas);
     children.forEach((c) => c.renderTree(canvas));
 
@@ -139,7 +139,7 @@ class Component with Loadable {
     if (debugMode) {
       renderDebugMode(canvas);
     }
-    postRender(canvas);
+    postRender(canvas); // ignore: deprecated_member_use_from_same_package
   }
 
   /// A render cycle callback that runs before the component and its children
@@ -148,12 +148,14 @@ class Component with Loadable {
   /// A possible use-case for this function is to transform the canvas, in which
   /// case you would also override `postRender()` in order to restore the state
   /// of the canvas as it was originally.
-  @protected
+  @nonVirtual
+  @Deprecated('Override renderTree() instead')
   void preRender(Canvas canvas) {}
 
   /// A render cycle callback that runs after the component and its children
   /// has been rendered.
-  @protected
+  @nonVirtual
+  @Deprecated('Override renderTree() instead')
   void postRender(Canvas canvas) {}
 
   void renderDebugMode(Canvas canvas) {}